check if Contestant is already in our database

+++ src/server/controller/api/contestant/contestantApiController.js
@@ -25,38 +25,56 @@ module.exports = class ContestantApiController {
       return response.status(400).json({success: false,
         error: {text: 'Es wurden nicht alle notwendingen Felder ausgefüllt'}});
     }
+    const contestantJSON = request.body;
 
-    StudentApiController.validate(request.body, (validated) => {
-      if (validated === true) {
-        const contestantJSON = request.body;
-        contestantJSON.activated = false;
-        contestantJSON.image = request.file.filename;
-        // sanitize user inputs
-        contestantJSON.firstName = xss(contestantJSON.firstName);
-        contestantJSON.lastName = xss(contestantJSON.lastName);
-        contestantJSON.course = xss(contestantJSON.course);
-        contestantJSON.year = xss(contestantJSON.year);
-        contestantJSON.centuria = '';
-        contestantJSON.description = xss(contestantJSON.description);
-        contestantJSON.token = '';
-        ContestantHelper.sendActivationMail(contestantJSON, (result) => {
-          if (result === false) {
-            return response.status(400).json({success: false,
-              error: {text: 'Fehler beim Versand der Bestätigungsmail'}});
-          }
-          return response.status(200).json({success: true});
-        });
-      } else if (validated === false && request.file.path !== undefined) {
-        fs.unlink(request.file.path, (error) => {
-          if (error) {
-            console.log(error);
-          }
-        });
-        return response.status(400).json({success: false,
-          error: {text: 'Deine Angaben konnten nicht validiert werden. \nVersuche es erneut'}});
-      }
-      // return next('error');
-    });
+    Contestant.count({firstName: {$regex: ContestantHelper.buildNameRegex(contestantJSON.firstName),
+      $options: 'g'},
+      lastName: contestantJSON.lastName}).exec((countError, count) => {
+        if (countError) {
+          return response.status(400).json({success: false,
+            error: {text: 'Deine Angaben konnten nicht validiert werden. \nVersuche es erneut'}});
+        }
+
+        if (count >= 1) {
+          return response.status(400).json({success: false,
+            error: {text: 'Du hast dich bereits auffgestellt.'}});
+        } else {
+          StudentApiController.validate(request.body, (validated) => {
+            if (validated === true) {
+              contestantJSON.activated = false;
+              contestantJSON.image = request.file.filename;
+              // sanitize user inputs
+              contestantJSON.firstName = xss(contestantJSON.firstName);
+              contestantJSON.lastName = xss(contestantJSON.lastName);
+              contestantJSON.course = xss(contestantJSON.course);
+              contestantJSON.year = xss(contestantJSON.year);
+              contestantJSON.centuria = '';
+              contestantJSON.description = xss(contestantJSON.description);
+              contestantJSON.token = '';
+              ContestantHelper.sendActivationMail(contestantJSON, (result) => {
+                if (result === false) {
+                  return response.status(400).json({
+                    success: false,
+                    error: {text: 'Fehler beim Versand der Bestätigungsmail'}
+                  });
+                }
+                return response.status(200).json({success: true});
+              });
+            } else if (validated === false && request.file.path !== undefined) {
+              fs.unlink(request.file.path, (error) => {
+                if (error) {
+                  console.log(error);
+                }
+              });
+              return response.status(400).json({
+                success: false,
+                error: {text: 'Deine Angaben konnten nicht validiert werden. \nVersuche es erneut'}
+              });
+            }
+            // return next('error');
+          });
+        }
+      });
   }
 
   static activate(request, response, next) {
